删除文件方法修改

+++ fhs_module/fhs_basics/fhs_basics_service/src/main/java/com/fhs/basics/service/impl/FileServerBusinessImpl.java
@@ -79,7 +79,9 @@ public class FileServerBusinessImpl implements FileServerBusiness {
     @Override
     public void deleteFile(String fileId) {
         PubFileVO pubFileVO = fileService.selectById(fileId);
-        fileStorage.deleteFile(pubFileVO);
+        if(pubFileVO != null){
+            fileStorage.deleteFile(pubFileVO);
+        }
         fileService.deletePubFile(fileId);
     }
